corrections

+++ app/Http/Controllers/Dsp/UsersController.php
@@ -16,9 +16,11 @@ use Vanguard\Http\Requests\User\UpdateUserRequest;
 use Vanguard\Services\User\UpdateService;
 use Vanguard\Services\User\ReinviteService;
 use Vanguard\User;
+use Vanguard\Models\Company;
 class UsersController extends Controller
 {
     use CompanyIdTrait;
+    const EMAIL_SUBJECT = "Invitation to join Vantage";
 
     public function __construct()
     {
@@ -35,8 +37,12 @@ class UsersController extends Controller
     {    
         $role_list_services = new ListRoleGroup('dsp');
         $roles = $role_list_services->getRoles();
+        $urls = ['list' => route('users.list'),
+                 'create' => route('users.invite')
+                ];
         return view('agency.user.index')
-               ->with('roles', $roles);
+               ->with('roles', $roles)
+               ->with('url', $urls);
     }
 
 
@@ -50,8 +56,9 @@ class UsersController extends Controller
     
     public function create(UserInviteRequest $request)
     {
+       
         $validated = $request->validated(); 
-        $invite_user = new InviteService($request, $this->companyId(), 'web');
+        $invite_user = new InviteService($validated, $this->companyId(), 'web', self::EMAIL_SUBJECT);
         $new_user = $invite_user->run();   
         return ['status'=>"success", 'message'=> "User(s) invited successfully, and emails sent"];
     }
@@ -59,12 +66,12 @@ class UsersController extends Controller
      /**
      * Return a list of user that the currently logged in user has permission to view
      */
-    public function list(ListRequest $request)
+    public function list()
     {        
-        $validated = $request->validated();
-        $validated['company_id'] = $this->companyId();   
-        $user_list_service = new UserListService($this->getCompanyIdsList());
-        $user_list = $user_list_service->run();
+        $user = \Auth::user();
+        $this->authorize('get', $user);
+        $company_user = Company::with('users')->findOrFail($this->companyId());
+        $user_list = $company_user->users;
         return new UserCollection($user_list);
     }
 
@@ -75,6 +82,8 @@ class UsersController extends Controller
     public function update(UpdateUserRequest $request, $id)
     {
         $validated = $request->validated(); 
+        $user = User::findOrFail($id);
+        $this->authorize('get', $user);
         $update_user_service = new UpdateService($validated, $this->companyId(), $id, 'web');
         $updated_user = $update_user_service->run();
         return new UserResource($updated_user);
@@ -83,8 +92,8 @@ class UsersController extends Controller
     public function resend($id)
     {
         $user = User::findOrFail($id);
-        $subject="Invitation to join Vantage";
-        $reinvite = new ReinviteService($user, $subject);
+        $this->authorize('get', $user);
+        $reinvite = new ReinviteService($user, self::EMAIL_SUBJECT);
         $user_reinvite = $reinvite->run();        
         return  $user_reinvite; 
     }
@@ -92,6 +101,7 @@ class UsersController extends Controller
     public function delete($id)
     {
         $user = User::findOrFail($id);
+        $this->authorize('get', $user);
         $user->delete();
         return ['status'=>"success", 'message'=> "User deleted successfully"];
     }
